add twitter login and ability to view tweets
This adds the ability to sign in via Twitter (using the device's social
media accounts) and view tweets. This is a very rough approach suitable
for the alpha release to show that this functionality is possible. There
needs to be more work on pulling down images (asynchronously) and
wrapping the text appropriately in the UICollectionView.