Tore out SwifteriOS and added Fabric for Twitter
This is a WIP commit. Not everything works yet, but the SwifteriOS
framework was torn out and replaced with Fabric. The login functionality
works as expected now but the pulling of tweets needs to be reworked.
Right now the project results in a build error in the main view
controller.